Re-organise GraphQL streaming tests in proper files (no functional changes)
